+++ ui/Monte_Carlo.py
@@ -37,12 +37,21 @@ else:
         st.info("Rates must be set to *stochastic* or *histochastic* to run Monte Carlo simulations.")
     else:
         st.markdown("Generate a histogram of results obtained from running multiple scenarios with stochastic rates.")
+        kz.initCaseKey("histogram_log_x", False)
         col1, col2, col3, col4 = st.columns(4, gap="large", vertical_alignment="bottom")
         with col1:
             kz.initCaseKey("MC_cases", 100)
             kz.getIntNum("Number of random instances", "MC_cases", step=10, max_value=10000)
         with col4:
             st.button("Run simulation", on_click=owb.runMC, disabled=kz.caseIsNotMCReady())
+
+        st.markdown("#####")
+        with st.expander("*Advanced options*"):
+            help_log_x = (
+                "Use log-spaced bins and a log-scale x-axis for the result histogram (log-normal style). "
+                "Values below $1k are excluded when on."
+            )
+            kz.getToggle("Log scale (x-axis) for histogram", "histogram_log_x", callback=kz.setpull, help=help_log_x)
 
     st.divider()
     fig = kz.getCaseKey("monteCarloPlot")